change notifir 4

+++ lib/models/weather_model.dart
@@ -1,6 +1,6 @@
 class WeatherModel {
   String date;
-  int temp;
+  double temp;
   double maxTemp;
   double minTemp;
   String WeatherstateName;
@@ -15,7 +15,6 @@ class WeatherModel {
 
   factory WeatherModel.fromJson(dynamic data) {
     var jsonData = data['forecast']['forecastday'][0]['day'];
-
 
     return WeatherModel(
       temp: jsonData['avgtemp_c'],
